refactor(core): colocate definition-time validation under src/validation/
Compile-time (branded FIX_* validators) and runtime definition checks
were scattered across types.ts, command/crust.ts, parsing/validation.ts,
and parsing/flag-validation.ts. Group them by kind so each rule's
type-level and runtime twin live side by side:

- validation/flags.ts: ValidateNamedFlagDefs, ValidateFlagAliases,
  ValidateNoPrefixedFlags + validateIncomingFlag
- validation/args.ts: ValidateVariadicArgs, AppendArgsChecks +
  validateSchemaExclusivity
- validation/commands.ts: validateAliasString, validateIncomingAliases
- validation/tree.ts: validateCommandTree, createValidationArgv

Pure move: no behavior change, public exports byte-identical, old
parsing/{validation,flag-validation}.ts deleted with no shims. Tests
moved co-located, plus new type-level coverage for FIX_ALIAS_COLLISION
and FIX_NO_PREFIX (previously untested). Parse-time input validation
stays in parsing/; AddChecks stays beside .add() to avoid a type cycle.

Also fixes the stale .args() signature in the API docs left over from
the accumulative-args refactor (#176).
