+++ packages/core/src/command/crust.ts
@@ -9,8 +9,6 @@ import type {
 } from "../api/context.ts";
 import type { Extension } from "../api/extension.ts";
 import { CrustError } from "../errors.ts";
-import { validateIncomingFlag } from "../parsing/flag-validation.ts";
-import { validateIncomingAliases } from "../parsing/validation.ts";
 import type {
 	ArgDef,
 	ArgsDef,
@@ -24,9 +22,10 @@ import type {
 	MergeFlags,
 	NamedFlagDef,
 	NamedFlagsRecord,
-	ValidateNamedFlagDefs,
-	ValidateVariadicArgs,
 } from "../types.ts";
+import { type AppendArgsChecks, validateSchemaExclusivity } from "../validation/args.ts";
+import { validateIncomingAliases } from "../validation/commands.ts";
+import { validateIncomingFlag, type ValidateNamedFlagDefs } from "../validation/flags.ts";
 import {
 	cloneCommandNode,
 	executeInvocation,
@@ -67,32 +66,6 @@ export interface CrustCommandContext<
 	command: CommandSnapshot;
 	/** Readonly snapshot of the application root, including Extension contributions */
 	rootCommand: CommandSnapshot;
-}
-
-// ────────────────────────────────────────────────────────────────────────────
-// Internal helpers — runtime flag validation
-// ────────────────────────────────────────────────────────────────────────────
-
-/**
- * Runtime guard for untyped callers: schema mode is exclusive — the schema
- * owns coercion, defaults, requiredness, choices, and validation.
- * The type system already rejects mixing; this catches plain-JS misuse.
- */
-function validateSchemaExclusivity(
-	subject: "arg" | "flag",
-	name: string,
-	def: Record<string, unknown>,
-): void {
-	if (def.schema === undefined) return;
-	for (const key of ["default", "required", "choices", "parse"] as const) {
-		if (def[key] !== undefined) {
-			throw new CrustError(
-				"DEFINITION",
-				`${subject} "${name}" mixes core option "${key}" with a schema — the schema exclusively owns coercion, defaults, requiredness, choices, and validation`,
-				{ subject, name, reason: "schema-exclusive" },
-			);
-		}
-	}
 }
 
 // ────────────────────────────────────────────────────────────────────────────
@@ -266,19 +239,6 @@ type DefinitionRequirements<D> = D extends CommandDefinition<infer R> ? R : neve
 type AppendedArgs<A extends ArgsDef, NewA extends ArgsDef> = ArgsDef extends A
 	? NewA
 	: readonly [...A, ...NewA];
-
-type AppendArgsChecks<A extends ArgsDef, NewA extends ArgsDef> = A extends readonly [
-	...unknown[],
-	infer Last,
-]
-	? Last extends { variadic: true }
-		? {
-				[I in keyof NewA]: NewA[I] & {
-					readonly FIX_VARIADIC_POSITION: "Only the last positional argument can be variadic";
-				};
-			}
-		: ValidateVariadicArgs<NewA>
-	: ValidateVariadicArgs<NewA>;
 
 /** Per-definition add-time checks (compile-time counterpart of the runtime attach checks). */
 type AddChecks<Ctx extends ContextMap, Ds extends readonly CommandDefinition<any>[]> = {
